renamed menu-controller to meals-controller
